requirements.txt
asgiref==3.7.1
Django==4.2.1
django-stdimage==6.0.1
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.1
